pyproject.toml
[project]
name = "dsa-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "gunicorn>=25.1.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.26.7",
    "pypdf>=6.5.0",
    "pyspellchecker>=0.9.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
]
